* origin/master: (32 commits)
  release(network-contracts): v6.1.0
  release(config): @streamr/config 4.0.5
  ETH-585 docker images build and test in CI (#616)
  Eth 596 subgraph fixeschanges (#631)
  fix: merge hazard from flagWithMetadata PR
  cleanup: reorder setParam function, uint256 -> uint in Policies
  feat: make slashing fraction configurable (#624)
  rename flagWithMetadata -> flag, and remove the previous implementation (#630)
  release(config): @streamr/config 4.0.4
  streamr 1.0 mumbai deployment addresses update
  Modularize operator (#629)
  doc: add clean step to push-to-docker-hub instructions
  release(network-contracts): 6.0.1
  fix: remove accidentally duplicated function
  Eth 584 erc677 in operator factory (#627)
  added hardhat-contract-sizer to contract build process
  ETH-579: Check Operator contract function guards (#628)
  deoc: added addresses preloaded with DATA to fastchain readme
  release(config): @streamr/config 4.0.3
  new addresses for mumbai streamr 1.0 deployment
  ...

+++ packages/network-contracts/contracts/OperatorTokenomics/SponsorshipPolicies/AdminKickPolicy.sol
@@ -15,7 +15,7 @@ contract AdminKickPolicy is IKickPolicy, Sponsorship {
         assembly {data.slot := storagePosition}
     }
 
-    function setParam(uint256 adminAdress) external {
+    function setParam(uint adminAdress) external {
         localData().admin = address(uint160(adminAdress));
     }
